requirements.txt
Pillow==9.4.0
tensorboard==2.12.0
tensorboardX==2.6
numpy==1.24.2
einops==0.6.0
torch-optimizer==0.3.0
tqdm==4.63.0
torch-summary==1.4.5
torchmetrics==0.11.3
black==22.8.0
piq==0.8.0
onnx==1.15.0